Create T-Test Analysis of Flipper Length in Male and Female Adelie Penguins
This script performs an independent t-test to compare the flipper lengths of male and female Adelie penguins using the Palmer Penguins dataset. It visualizes the results with a bar plot and includes the t-statistic and p-value in the title. This analysis helps determine whether there is a statistically significant difference in flipper length between the two sexes.